requirements.txt
discord.py
python-dotenv
aiohttp
pytesseract
Pillow



